+++ lib/disparity_graph.cpp
@@ -22,7 +22,9 @@
  * SOFTWARE.
  */
 #include <disparity_graph.hpp>
+
 #include <stdexcept>
+#include <string>
 
 #define SQR(x) ((x) * (x))
 #define TO_FLOAT(x) (static_cast<FLOAT>(x))
@@ -38,15 +40,11 @@ DisparityGraph::DisparityGraph(
 {
     if (!image_valid(this->left))
     {
-        throw std::invalid_argument(
-            "Left image is invalid."
-        );
+        throw std::invalid_argument("Left image is invalid.");
     }
     if (!image_valid(this->right))
     {
-        throw std::invalid_argument(
-            "Right image is invalid."
-        );
+        throw std::invalid_argument("Right image is invalid.");
     }
     if (this->maximal_disparity == 0)
     {
@@ -57,27 +55,59 @@ DisparityGraph::DisparityGraph(
     if (this->maximal_disparity >= this->left.width)
     {
         throw std::invalid_argument(
-            "Maximal disparity should be less than width of the left image."
+            "Maximal disparity should be less than width of the left image. "
+            "Width of the left image is "
+            + std::to_string(this->left.width)
+            + ". Provided maximal disparity is "
+            + std::to_string(this->maximal_disparity)
+            + "."
         );
     }
     if (this->left.width != this->right.width)
     {
         throw std::invalid_argument(
-            "Number of columns of the images should be equal."
+            "Number of columns of the images should be equal. "
+            "Current left and right images have "
+            + std::to_string(this->left.width)
+            + " and "
+            + std::to_string(this->right.width)
+            + " columns respectively."
         );
     }
     if (this->left.height != this->right.height)
     {
         throw std::invalid_argument(
-            "Number of rows of the images should be equal."
+            "Number of rows of the images should be equal. "
+            "Current left and right images have "
+            + std::to_string(this->left.height)
+            + " and "
+            + std::to_string(this->right.height)
+            + " rows respectively."
         );
     }
     if (this->left.max_value != this->right.max_value)
     {
         throw std::invalid_argument(
-            "Maximal intensity of the images should be the same."
+            "Maximal intensity of the images should be the same. "
+            "Maximal intensity of provided left and right images "
+            + std::to_string(this->left.max_value)
+            + " and "
+            + std::to_string(this->right.max_value)
+            + "respectively."
         );
     }
+
+    this->reparametrization.resize(
+        this->left.width
+        * this->left.height
+        * NEIGHBORS_COUNT
+        * this->maximal_disparity
+    );
+    fill(
+        this->reparametrization.begin(),
+        this->reparametrization.end(),
+        static_cast<FLOAT>(0)
+    );
 }
 
 ULONG neighbor_index(
@@ -189,10 +219,84 @@ bool edge_exists(
     return true;
 }
 
-FLOAT edge_penalty(struct Edge edge)
+ULONG reparametrization_index_fast(
+    const struct DisparityGraph& graph,
+    struct Node node,
+    ULONG neighbor_index
+)
+{
+    ULONG index = 0;
+    index *= graph.right.width;
+    index += node.pixel.column;
+    index *= NEIGHBORS_COUNT;
+    index += neighbor_index;
+    index *= graph.maximal_disparity;
+    index += node.disparity;
+    index *= graph.left.height;
+    index += node.pixel.row;
+
+    return index;
+}
+
+ULONG reparametrization_index(
+    const struct DisparityGraph& graph,
+    struct Node node,
+    struct Pixel neighbor
+)
+{
+    return reparametrization_index_fast(
+        graph,
+        node,
+        neighbor_index(node.pixel, neighbor)
+    );
+}
+
+ULONG reparametrization_index_slow(
+    const struct DisparityGraph& graph,
+    struct Edge edge
+)
+{
+    return reparametrization_index_fast(
+        graph,
+        edge.node,
+        neighbor_index(edge.node.pixel, edge.neighbor.pixel)
+    );
+}
+
+FLOAT reparametrization_value(
+    const struct DisparityGraph& graph,
+    struct Node node,
+    struct Pixel neighbor
+)
+{
+    return graph.reparametrization[reparametrization_index(graph, node, neighbor)];
+}
+
+FLOAT reparametrization_value_slow(
+    const struct DisparityGraph& graph,
+    struct Edge edge
+)
+{
+    return graph.reparametrization[reparametrization_index_slow(graph, edge)];
+}
+
+FLOAT reparametrization_value_fast(
+    const struct DisparityGraph& graph,
+    struct Node node,
+    ULONG neighbor_index
+)
+{
+    return graph.reparametrization[
+        reparametrization_index_fast(graph, node, neighbor_index)
+    ];
+}
+
+FLOAT edge_penalty(const struct DisparityGraph& graph, struct Edge edge)
 {
     return
-        SQR(TO_FLOAT(edge.node.disparity) - TO_FLOAT(edge.neighbor.disparity));
+        SQR(TO_FLOAT(edge.node.disparity) - TO_FLOAT(edge.neighbor.disparity))
+        + reparametrization_value_slow(graph, edge)
+        + reparametrization_value(graph, edge.neighbor, edge.node.pixel);
 }
 
 FLOAT node_penalty(const struct DisparityGraph& graph, struct Node node)
@@ -201,5 +305,9 @@ FLOAT node_penalty(const struct DisparityGraph& graph, struct Node node)
     left_pixel.column += node.disparity;
     return
         SQR(TO_FLOAT(get_pixel_value(graph.right, node.pixel))
-            - TO_FLOAT(get_pixel_value(graph.left, left_pixel)));
+            - TO_FLOAT(get_pixel_value(graph.left, left_pixel)))
+        - reparametrization_value_fast(graph, node, 0)
+        - reparametrization_value_fast(graph, node, 1)
+        - reparametrization_value_fast(graph, node, 2)
+        - reparametrization_value_fast(graph, node, 3);
 }