+++ lib/main.cpp
@@ -6,6 +6,7 @@
 #include <stdexcept>
 #include <string>
 
+#include <disparity_graph.hpp>
 #include <image.hpp>
 #include <pgm_io.hpp>
 
@@ -21,14 +22,28 @@ int main(int argc, char* argv[]) try
          "Left image")
         ("right-image,r",
          boost::program_options::value<std::string>(),
-         "Right image");
+         "Right image")
+        ("maximal-disparity,d",
+         boost::program_options::value<std::string>(),
+         "Maximal disparity");
 
     boost::program_options::variables_map vm;
-    boost::program_options::store(
-        boost::program_options::parse_command_line(argc, argv, desc),
-        vm
-    );
-    boost::program_options::notify(vm);   
+    try
+    {
+        boost::program_options::store(
+            boost::program_options::parse_command_line(argc, argv, desc),
+            vm
+        );
+    }
+    catch (boost::program_options::multiple_occurrences& e)
+    {
+        std::cerr
+            << "You should specify each parameter only once: "
+            << e.what() << std::endl;
+        return 1;
+    }
+
+    boost::program_options::notify(vm);
 
     if (vm.count("left-image") == 1 && vm.count("right-image") == 1)
     {
@@ -39,6 +54,22 @@ int main(int argc, char* argv[]) try
             };
             struct Image right_image{
                 read_image(vm["right-image"].as<std::string>())
+            };
+            ULONG maximal_disparity = 0;
+            if (vm.count("maximal-disparity") == 0)
+            {
+                maximal_disparity = left_image.width - 1;
+            }
+            else
+            {
+                maximal_disparity = std::stoul(
+                    vm["maximal-disparity"].as<std::string>()
+                );
+            }
+            struct DisparityGraph disparity_graph{
+                left_image,
+                right_image,
+                maximal_disparity
             };
         }
         catch (std::invalid_argument& e)