+++ tests/disparity_graph.cpp
@@ -180,34 +180,34 @@ BOOST_AUTO_TEST_CASE(check_initial_edges_penalties)
     struct DisparityGraph disparity_graph{left_image, right_image, 2};
 
     BOOST_CHECK_CLOSE(
-        edge_penalty({{{0, 0}, 0}, {{0, 1}, 0}}),
+        edge_penalty(disparity_graph, {{{0, 0}, 0}, {{0, 1}, 0}}),
         0.0,
         0.5
     );
     BOOST_CHECK_CLOSE(
-        edge_penalty({{{0, 0}, 0}, {{1, 0}, 1}}),
+        edge_penalty(disparity_graph, {{{0, 0}, 0}, {{1, 0}, 1}}),
         1.0,
         0.5
     );
 
     BOOST_CHECK_CLOSE(
-        edge_penalty({{{0, 1}, 1}, {{0, 2}, 0}}),
+        edge_penalty(disparity_graph, {{{0, 1}, 1}, {{0, 2}, 0}}),
         1.0,
         0.5
     );
 
     BOOST_CHECK_CLOSE(
-        edge_penalty({{{0, 1}, 0}, {{0, 0}, 2}}),
+        edge_penalty(disparity_graph, {{{0, 1}, 0}, {{0, 0}, 2}}),
         4.0,
         0.5
     );
     BOOST_CHECK_CLOSE(
-        edge_penalty({{{0, 1}, 2}, {{1, 0}, 0}}),
+        edge_penalty(disparity_graph, {{{0, 1}, 2}, {{1, 0}, 0}}),
         4.0,
         0.5
     );
     BOOST_CHECK_CLOSE(
-        edge_penalty({{{0, 1}, 2}, {{1, 1}, 2}}),
+        edge_penalty(disparity_graph, {{{0, 1}, 2}, {{1, 1}, 2}}),
         0.0,
         0.5
     );
@@ -258,6 +258,99 @@ BOOST_AUTO_TEST_CASE(check_initial_nodes_penalties)
     BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{1, 1}, 1}), 1.0, 0.5);
 
     BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{1, 2}, 0}), 0.0, 0.5);
+}
+
+BOOST_AUTO_TEST_CASE(check_edges_penalties)
+{
+    PGM_IO pgm_io;
+    std::istringstream left_image_content{R"left_image(
+    P2
+    3 2
+    2
+    0 1 2
+    2 0 1
+    )left_image"};
+    std::istringstream right_image_content{R"right_image(
+    P2
+    3 2
+    2
+    1 0 2
+    2 0 1
+    )right_image"};
+
+    left_image_content >> pgm_io;
+    BOOST_REQUIRE(pgm_io.get_image());
+    struct Image left_image{*pgm_io.get_image()};
+
+    right_image_content >> pgm_io;
+    BOOST_REQUIRE(pgm_io.get_image());
+    struct Image right_image{*pgm_io.get_image()};
+
+    struct DisparityGraph disparity_graph{left_image, right_image, 2};
+
+    disparity_graph.reparametrization[
+        reparametrization_index(disparity_graph, {{0, 0}, 0}, {0, 1})
+    ] = -2.0;
+
+    BOOST_CHECK_CLOSE(
+        edge_penalty(disparity_graph, {{{0, 0}, 0}, {{0, 1}, 0}}),
+        -2.0,
+        0.5
+    );
+    BOOST_CHECK_CLOSE(
+        edge_penalty(disparity_graph, {{{0, 0}, 0}, {{0, 1}, 1}}),
+        -1.0,
+        0.5
+    );
+
+    BOOST_CHECK_CLOSE(
+        edge_penalty(disparity_graph, {{{0, 1}, 0}, {{0, 0}, 0}}),
+        -2.0,
+        0.5
+    );
+    BOOST_CHECK_CLOSE(
+        edge_penalty(disparity_graph, {{{0, 1}, 0}, {{0, 0}, 1}}),
+        1.0,
+        0.5
+    );
+}
+
+BOOST_AUTO_TEST_CASE(check_nodes_penalties)
+{
+    PGM_IO pgm_io;
+    std::istringstream left_image_content{R"left_image(
+    P2
+    3 2
+    2
+    0 1 2
+    2 0 1
+    )left_image"};
+    std::istringstream right_image_content{R"right_image(
+    P2
+    3 2
+    2
+    1 0 2
+    2 0 1
+    )right_image"};
+
+    left_image_content >> pgm_io;
+    BOOST_REQUIRE(pgm_io.get_image());
+    struct Image left_image{*pgm_io.get_image()};
+
+    right_image_content >> pgm_io;
+    BOOST_REQUIRE(pgm_io.get_image());
+    struct Image right_image{*pgm_io.get_image()};
+
+    struct DisparityGraph disparity_graph{left_image, right_image, 2};
+
+    disparity_graph.reparametrization[
+        reparametrization_index(disparity_graph, {{0, 0}, 0}, {0, 1})
+    ] = -2.0;
+
+    BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{0, 0}, 0}), 3.0, 0.5);
+    BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{0, 0}, 1}), 0.0, 0.5);
+
+    BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{0, 1}, 0}), 1.0, 0.5);
 }
 
 BOOST_AUTO_TEST_SUITE_END()